requirements.txt
pydantic-settings==2.12.0
pytest-dotenv==0.5.2
openai==2.8.1
sentence-transformers==5.1.2
torch==2.9.1
chromadb==1.3.5
pypdf==6.4.0
requests==2.32.5
